requirements.txt
librosa==0.9.1
faster-whisper==0.9.0
pydub==0.25.1
wavmark
numpy==1.22.0
eng_to_ipa==0.0.2
inflect==7.0.0
unidecode==1.3.7
whisper-timestamped==1.14.2
openai
python-dotenv
pypinyin==0.50.0
cn2an==0.5.22
jieba==0.42.1
langid==1.1.6
accelerate==0.25.0
huggingface-hub==0.20.1
Pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
einops==0.7.0
gradio==4.15.0
huggingface_hub
ffmpeg-python
moviepy
opencv-python
faster_whisper
openai-whisper
pyaudio
SpeechRecognition